requirements.txt
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pypdf>=6.0.0
arxiv>=1.4.7
schedule>=1.2.0
email-validator>=2.1.0
pyyaml>=6.0
numpy>=1.24.0
aiohttp>=3.8.0